Improved Networking by adding a NetworkObject that gets sent through the network. This makes it easy to compact data later on if required.
Changed ClientListener.sendData to contain a NetwortObject type instead
of object. This allows better navigation of the object when a client
recieves information.

Added small tree parser that allows commands fomr the host

Once a client chanegs their name, it is then changed on the server and
alerts every client of the change.