Examples of Packet acknowledgement have been implemented in ChatRoom and Chat Client.
Refracted the NetworkObject class to include a type of NetworkData
instead of plainly using Object. This will allow for seperating
ChatMessages and Moves easier when merging with Game Logic.

GameServer and Game Client classes made for implementation on the Gui
and GameLogic sides. Move and ChatMessage classes have been made for
both classes as well so clients can send messages OR perform a move on
the server.