Synchronized the ChatClient more.
Chat Client now has a modified field instead of a date object to
determine when anything is changed in the messages.

Improved the chat system in ChatRoom to work better with the ChatClient

Messages now display with a > to indicate if the message is modified or
not. This is updatated when the message is sent back to the client after
sending.

Added /clear command to the clients ( never gets sent to server )