Incoorporated the GameServer to call "moveCharacterTo(move)" from game logic to start testing basic moves from the client side to the server side game logic.
Created the Player class to replace various parts of the Networking.
Contains IPAddress and Name of the player according to the client and
server side.

RoomUpdate is a wrapped object of the updated room sent from the
GameLogic to the Server. Where it updates the clients version of the
room.

Incoorporated Color for chat messages so they may now have color when
drawn on the screen.