Avatar - added fields to indicate the center coordinates for Tiles. created centerAvatar() - updates avatars coordinates to be in the center of its current tile. Avatar now contains a Color color, field which can be changed when it interacts with a ColorChange object, and hopefully soon to unlock same color doors.
Charger - created inner class in charger that extends Thread. When an
avatar interacts with a charger object, a new thread is created which
sets the avatars charging field to true iff / as long as the avatar
remains on the same tile as the charger. when the avatar leaves the
tile, the thread sets the charging field to false.

Door - now Contain Color color fields, the color of the door and the
color of the 'key' needed to access it.

Item - added abstract method - pickItemUp().

ColorChange - new class. extends Item. idea is to represent a system
that changes the color of the avatar that interacts with it.