Avatar can now not move through a door if they doesnt have the relative key  color in their inventory.
Bug with picking up items - coulnt pick up an item if their was 4 items
in it, fixed so that the whole inventory can be filled (5 items)

updated NewGame class that creates the doors and sets the links between
rooms. Doors are now created as one of the 4 colors and match on both
rooms they are attached to. Colored doors can be passed through if the
avatar has the same color key in their inventory.